+++ backend/src/dao/state_dao.rs
@@ -2,7 +2,7 @@ use super::Table;
 use crate::model::geo_id;
 
 impl<'c> Table<'c, geo_id::State> {
-    pub async fn by_id(&self, id: i32) -> Result<geo_id::State, sqlx::Error> {
+    pub async fn by_id(&self, id: i64) -> Result<geo_id::State, sqlx::Error> {
         sqlx::query_as!(
             geo_id::State,
             "
